PORTALS-3582
Lazy-load routes in CCKP
