+++ apps/portals/cancercomplexity/src/config/paletteConfig.ts
@@ -1,4 +1,4 @@
-import { Palettes } from 'synapse-react-client'
+import Palettes from 'synapse-react-client/theme/palette/Palettes'
 
 const palette = Palettes.cancerComplexityPortalPalette
