+++ apps/portals/cancercomplexity/src/config/routesConfig.tsx
@@ -1,38 +1,8 @@
 import App from '@sage-bionetworks/synapse-portal-framework/App'
-import AboutPortal from '@sage-bionetworks/synapse-portal-framework/components/csbc-home-page/AboutPortal'
-import ConsortiaGoals from '@sage-bionetworks/synapse-portal-framework/components/csbc-home-page/ConsortiaGoals'
-import DevelopedBySage from '@sage-bionetworks/synapse-portal-framework/components/csbc-home-page/DevelopedBySage'
-import Ecosystem from '@sage-bionetworks/synapse-portal-framework/components/csbc-home-page/Ecosystem'
-import ExploreWrapper from '@sage-bionetworks/synapse-portal-framework/components/Explore/ExploreWrapper'
-import { SectionLayout } from '@sage-bionetworks/synapse-portal-framework/components/SectionLayout'
-import sharedRoutes from '@sage-bionetworks/synapse-portal-framework/shared-config/sharedRoutes'
-import { RouteObject } from 'react-router'
-import {
-  RssFeedCards,
-  ThemesPlot,
-  Programs,
-} from 'synapse-react-client'
-import DatasetsDetailsPage from '../pages/DatasetsDetailsPage'
-import EducationalResourcesDetailsPage from '../pages/EducationalResourcesDetailsPage'
-import GrantDetailsPage from '../pages/GrantDetailsPage'
-import PeopleDetailsPage from '../pages/PeopleDetailsPage'
-import PublicationsDetailsPage from '../pages/PublicationsDetailsPage'
-import ToolsDetailsPage from '../pages/ToolsDetailsPage'
-import explorePageRoutes from './explorePageRoutes'
-import {
-  consortiaHomePageConfig,
-} from './synapseConfigs/consortiaHomePage'
-import { MC2Supplement } from './synapseConfigs/MC2Supplement'
-import {
-  onIndividualThemeBarPlotPointClick,
-  onPointClick,
-} from './synapseConfigs/onPointClick'
-import { searchPageChildRoutes } from '@/pages/CCKPSearchPage'
-import CancerComplexityHeader from '@sage-bionetworks/synapse-portal-framework/components/cancercomplexity/CancerComplexityHeader'
-import CancerComplexityIntro from '@sage-bionetworks/synapse-portal-framework/components/cancercomplexity/CancerComplexityIntro'
 import RepositoryUnderReviewAlert from '@sage-bionetworks/synapse-portal-framework/components/RepositoryUnderReviewAlert'
-import { SharePageLinkButton } from 'synapse-react-client'
-import { sharePageLinkExplorePageButtonProps } from '@sage-bionetworks/synapse-portal-framework/shared-config/SharePageLinkButtonConfig'
+import sharedRoutes from '@sage-bionetworks/synapse-portal-framework/shared-config/sharedRoutes'
+import { convertModuleToRouteObject } from '@sage-bionetworks/synapse-portal-framework/utils/convertModuleToRouteObject'
+import { RouteObject } from 'react-router'
 
 const routes: RouteObject[] = [
   {
@@ -51,231 +21,106 @@ const routes: RouteObject[] = [
       ...sharedRoutes,
       {
         index: true,
-        element: (
-          <>
-            <CancerComplexityHeader />
-            <SectionLayout>
-              <CancerComplexityIntro />
-            </SectionLayout>
-            <div className={'home-bg-dark'}>
-              <SectionLayout
-                title={'Portal Goals'}
-                centerTitle
-                ContainerProps={{
-                  className: 'home-spacer',
-                }}
-              >
-                <ConsortiaGoals />
-              </SectionLayout>
-            </div>
-            {/* <SectionLayout
-              title="What Resources are Available?"
-              centerTitle
-              ContainerProps={{
-                className: 'home-spacer',
-              }}
-            >
-              <CardDeck {...availableResearchCardDeckConfig} />
-            </SectionLayout> */}
-
-            <SectionLayout
-              title={
-                'What Research Themes are Scientists Currently Focusing On?'
-              }
-              centerTitle
-              ContainerProps={{
-                className: 'home-spacer',
-              }}
-            >
-              <ThemesPlot
-                onPointClick={onPointClick}
-                onIndividualThemeBarPlotPointClick={
-                  onIndividualThemeBarPlotPointClick
-                }
-                topBarPlot={{
-                  entityId: 'syn21641485',
-                  xField: 'totalCount',
-                  yField: 'groupBy',
-                  groupField: 'consortium',
-                  colors: {
-                    CCBIR: 'rgba(24, 115, 107, 1)',
-                    CSBC: 'rgba(156, 196, 233, 1)',
-                    ICBP: 'rgba(226, 149, 12, 1)',
-                    MetNet: 'rgba(30, 44, 72, 1)',
-                    NCI: 'rgba(249, 204, 125, 1)',
-                    PDMC: 'rgba(32, 171, 159, 1)',
-                    'PS-ON': 'rgba(197, 191, 223, 1)',
-                    TEC: 'rgba(59, 88, 143, 1)',
-                    'Sage Bionetworks': 'rgba(127, 127, 127, 1)',
-                  },
-                  whereClause: 'totalCount is not NULL',
-                }}
-                sideBarPlot={{
-                  entityId: 'syn21649281',
-                  xField: 'totalCount',
-                  yField: 'theme',
-                  groupField: 'consortium',
-                  countLabel: 'grants',
-                  plotStyle: {
-                    backgroundColor: '#f5f9fa',
-                  },
-                  colors: {
-                    CSBC: '#1c76af',
-                    'PS-ON': '#5bb0b5',
-                    ICBP: '#9cc4e9',
-                    TEC: '#c5bfdf',
-                  },
-                }}
-                dotPlot={{
-                  entityId: 'syn21639584',
-                  xField: 'totalCount',
-                  yField: 'theme',
-                  groupField: 'groupBy',
-                  infoField: 'themeDescription',
-                  whereClause:
-                    "groupBy IN ('publications', 'tools', 'datasets')",
-                  markerSymbols: {
-                    tools: 'y-down',
-                    datasets: 'diamond-x',
-                    publications: 'circle',
-                  },
-                  plotStyle: {
-                    markerLine: 'rgba(0, 0, 0,1)',
-                    markerFill: 'rgba(255, 255, 255,1)',
-                    markerSize: 11,
-                    backgroundColor: '#f5f9fa',
-                  },
-                }}
-              />
-            </SectionLayout>
-            <div className={'home-bg-dark'}>
-              <SectionLayout
-                title={"What's New"}
-                centerTitle
-                ContainerProps={{ className: 'home-spacer' }}
-              >
-                <RssFeedCards
-                  url={'https://news.cancercomplexity.synapse.org'}
-                  itemsToShow={3}
-                  allowCategories={[]}
-                  filter={{
-                    value: 'CCKP',
-                    type: 'category',
-                  }}
-                />
-              </SectionLayout>
-            </div>
-            <SectionLayout
-              title={'Consortia'}
-              centerTitle
-              ContainerProps={{
-                className: 'home-spacer cancercomplexity-consortia',
-              }}
-            >
-              <Programs {...consortiaHomePageConfig} />
-            </SectionLayout>
-            <div className={'home-bg-dark'}>
-              <SectionLayout
-                title="The Cancer Resource Information Ecosystem"
-                subtitle="The Cancer Resource Information ecosystem contains a growing list of tools and resources. Explore some of them below."
-                centerTitle
-                ContainerProps={{
-                  className: 'home-spacer',
-                }}
-              >
-                <Ecosystem
-                  config={[
-                    {
-                      title: 'Public Repositories',
-                      ownerId: 'syn21498902',
-                      wikiId: '601489',
-                    },
-                    {
-                      title: 'Data Commons',
-                      ownerId: 'syn21498902',
-                      wikiId: '601490',
-                    },
-                    {
-                      title: 'Data Coordination',
-                      ownerId: 'syn21498902',
-                      wikiId: '601574',
-                    },
-                    {
-                      title: 'Data Exploration',
-                      ownerId: 'syn21498902',
-                      wikiId: '601575',
-                    },
-                    {
-                      title: 'Cloud Platform',
-                      ownerId: 'syn21498902',
-                      wikiId: '601576',
-                    },
-                  ]}
-                />
-              </SectionLayout>
-            </div>
-            <SectionLayout
-              title={'About the Cancer Complexity Knowledge Portal'}
-              subtitle={
-                'The portal is built to disseminate resources to accelerate discovery and collaboration in the cancer research community. We aim to provide rich context about and access to activities and contributors that have produced the resources hosted within this and other repositories.'
-              }
-              centerTitle
-              ContainerProps={{ className: 'home-spacer' }}
-            >
-              <AboutPortal
-                cardProps={[
-                  { ownerId: 'syn21498902', wikiId: '601369' },
-                  { ownerId: 'syn21498902', wikiId: '601370' },
-                ]}
-              />
-            </SectionLayout>
-            <SectionLayout>
-              <DevelopedBySage />
-            </SectionLayout>
-          </>
-        ),
+        lazy: () => import('@/pages/Home').then(convertModuleToRouteObject),
       },
       {
         path: 'Explore',
-        element: (
-          <>
-            <SharePageLinkButton {...sharePageLinkExplorePageButtonProps} />
-            <ExploreWrapper explorePaths={explorePageRoutes} />
-          </>
-        ),
-        children: explorePageRoutes,
+        lazy: () =>
+          import('@/pages/Explore/layout').then(convertModuleToRouteObject),
+        children: [
+          {
+            path: 'Grants',
+            lazy: () =>
+              import('@/pages/Explore/Grants').then(convertModuleToRouteObject),
+          },
+          {
+            path: 'People',
+            lazy: () =>
+              import('@/pages/Explore/People').then(convertModuleToRouteObject),
+          },
+          {
+            path: 'Publications',
+            lazy: () =>
+              import('@/pages/Explore/Publications').then(
+                convertModuleToRouteObject,
+              ),
+          },
+          {
+            path: 'Datasets',
+            lazy: () =>
+              import('@/pages/Explore/Datasets').then(
+                convertModuleToRouteObject,
+              ),
+          },
+          {
+            path: 'Tools',
+            lazy: () =>
+              import('@/pages/Explore/Tools').then(convertModuleToRouteObject),
+          },
+          {
+            path: 'Educational Resources',
+            lazy: () =>
+              import('@/pages/Explore/EducationalResources').then(
+                convertModuleToRouteObject,
+              ),
+          },
+        ],
       },
       {
         path: 'Explore/Grants/DetailsPage',
-        element: <GrantDetailsPage />,
+        lazy: () =>
+          import('@/pages/GrantDetailsPage').then(convertModuleToRouteObject),
       },
       {
         path: 'Explore/People/DetailsPage',
-        element: <PeopleDetailsPage />,
+        lazy: () =>
+          import('@/pages/PeopleDetailsPage').then(convertModuleToRouteObject),
       },
       {
         path: 'Explore/Publications/DetailsPage',
-        element: <PublicationsDetailsPage />,
+        lazy: () =>
+          import('@/pages/PublicationsDetailsPage').then(
+            convertModuleToRouteObject,
+          ),
       },
       {
         path: 'Explore/Datasets/DetailsPage',
-        element: <DatasetsDetailsPage />,
+        lazy: () =>
+          import('@/pages/DatasetsDetailsPage').then(
+            convertModuleToRouteObject,
+          ),
       },
       {
         path: 'Explore/Tools/DetailsPage',
-        element: <ToolsDetailsPage />,
+        lazy: () =>
+          import('@/pages/ToolsDetailsPage').then(convertModuleToRouteObject),
       },
       {
         path: 'Explore/Educational Resources/DetailsPage',
-        element: <EducationalResourcesDetailsPage />,
+        lazy: () =>
+          import('@/pages/EducationalResourcesDetailsPage').then(
+            convertModuleToRouteObject,
+          ),
       },
       {
         path: 'MC2Supplement',
-        element: <MC2Supplement />,
+        lazy: () =>
+          import('@/pages/MC2Supplement').then(convertModuleToRouteObject),
       },
       {
         path: 'Search',
-        children: searchPageChildRoutes,
+        children: [
+          {
+            index: true,
+            lazy: () =>
+              import('@/pages/Search').then(convertModuleToRouteObject),
+          },
+          {
+            path: ':resourceType',
+            lazy: () =>
+              import('@/pages/Search').then(convertModuleToRouteObject),
+          },
+        ],
       },
     ],
   },
