+++ apps/portals/cancercomplexity/src/config/synapseConfigs/people.ts
@@ -1,9 +1,9 @@
 import personSvg from '@/config/style/Person.svg?url'
-import {
+import type {
   CardConfiguration,
   QueryWrapperPlotNavProps,
-  SynapseConstants,
 } from 'synapse-react-client'
+import * as SynapseConstants from 'synapse-react-client/utils/SynapseConstants'
 import { TableToGenericCardMapping } from 'synapse-react-client/components/GenericCard/TableRowGenericCard'
 import { peopleSql } from '../resources'
