+++ apps/portals/cancercomplexity/src/pages/Explore/Datasets.tsx
@@ -1,0 +1,7 @@
+import { datasetsQueryWrapperPlotNavProps } from '@/config/synapseConfigs'
+import { QueryWrapperPlotNav } from 'synapse-react-client/components/QueryWrapperPlotNav'
+
+function ExploreDatasets() {
+  return <QueryWrapperPlotNav {...datasetsQueryWrapperPlotNavProps} />
+}
+export default ExploreDatasets
