+++ apps/portals/cancercomplexity/src/pages/Explore/EducationalResources.tsx
@@ -1,0 +1,8 @@
+import { educationQueryWrapperPlotNavProps } from '@/config/synapseConfigs'
+import { QueryWrapperPlotNav } from 'synapse-react-client/components/QueryWrapperPlotNav'
+
+function ExploreEducationalResources() {
+  return <QueryWrapperPlotNav {...educationQueryWrapperPlotNavProps} />
+}
+
+export default ExploreEducationalResources
