+++ apps/portals/cancercomplexity/src/pages/Explore/Grants.tsx
@@ -1,0 +1,7 @@
+import { grantQueryWrapperPlotNavProps } from '@/config/synapseConfigs'
+import { QueryWrapperPlotNav } from 'synapse-react-client/components/QueryWrapperPlotNav'
+
+function ExploreGrants() {
+  return <QueryWrapperPlotNav {...grantQueryWrapperPlotNavProps} />
+}
+export default ExploreGrants
