+++ apps/portals/cancercomplexity/src/pages/Explore/People.tsx
@@ -1,0 +1,7 @@
+import { peopleQueryWrapperPlotNavProps } from '@/config/synapseConfigs'
+import { QueryWrapperPlotNav } from 'synapse-react-client/components/QueryWrapperPlotNav'
+
+function ExplorePeople() {
+  return <QueryWrapperPlotNav {...peopleQueryWrapperPlotNavProps} />
+}
+export default ExplorePeople
