+++ apps/portals/cancercomplexity/src/pages/Explore/Publications.tsx
@@ -1,0 +1,7 @@
+import { publicationsQueryWrapperPlotNavProps } from '@/config/synapseConfigs'
+import { QueryWrapperPlotNav } from 'synapse-react-client/components/QueryWrapperPlotNav'
+
+function ExplorePublications() {
+  return <QueryWrapperPlotNav {...publicationsQueryWrapperPlotNavProps} />
+}
+export default ExplorePublications
