+++ apps/portals/cancercomplexity/src/pages/Explore/Tools.tsx
@@ -1,0 +1,7 @@
+import { toolsQueryWrapperPlotNavProps } from '@/config/synapseConfigs'
+import { QueryWrapperPlotNav } from 'synapse-react-client/components/QueryWrapperPlotNav'
+
+function ExploreTools() {
+  return <QueryWrapperPlotNav {...toolsQueryWrapperPlotNavProps} />
+}
+export default ExploreTools
