+++ apps/portals/cancercomplexity/src/pages/ToolsDetailsPage.tsx
@@ -3,11 +3,6 @@ import { DetailsPageContextConsumer } from '@sage-bionetworks/synapse-portal-fra
 import DetailsPage from '@sage-bionetworks/synapse-portal-framework/components/DetailsPage/index'
 import { useGetPortalComponentSearchParams } from '@sage-bionetworks/synapse-portal-framework/utils/UseGetPortalComponentSearchParams'
 import { ColumnSingleValueFilterOperator } from '@sage-bionetworks/synapse-types'
-import {
-  CardContainerLogic,
-  ErrorPage,
-  SynapseErrorType,
-} from 'synapse-react-client'
 import columnAliases from '../config/columnAliases'
 import {
   grantsSql,
@@ -20,10 +15,14 @@ import { grantsCardConfiguration } from '../config/synapseConfigs/grants'
 import { peopleCardConfiguration } from '../config/synapseConfigs/people'
 import { publicationsCardConfiguration } from '../config/synapseConfigs/publications'
 import { toolsConfiguration } from '../config/synapseConfigs/tools'
-import { SharePageLinkButton } from 'synapse-react-client'
+import { SharePageLinkButton } from 'synapse-react-client/components/SharePageLinkButton'
 import { sharePageLinkButtonDetailPageProps } from '@sage-bionetworks/synapse-portal-framework/shared-config/SharePageLinkButtonConfig'
+import ErrorPage, {
+  SynapseErrorType,
+} from 'synapse-react-client/components/error/ErrorPage'
+import CardContainerLogic from 'synapse-react-client/components/CardContainerLogic'
 
-export default function ToolsDetailsPage() {
+function ToolsDetailsPage() {
   const { toolName } = useGetPortalComponentSearchParams()
 
   if (!toolName) {
@@ -118,3 +117,5 @@ export default function ToolsDetailsPage() {
     </>
   )
 }
+
+export default ToolsDetailsPage
